+++ deep_causality/src/types/reasoning_types/causable/mod.rs
@@ -3,3 +3,4 @@
 
 pub mod causaloid;
 pub mod causaloid_graph;
+mod causaloid_utils;
